refactor: delete vesta-common, inline code into CLI and vestad
Each crate now owns its code directly:
- CLI: client.rs (HTTP + WS), common.rs (types, config, helpers)
- vestad: types.rs (BackupInfo/BackupType), update_check.rs (inlined)
- tests: duplicated client + types for integration tests
