+++ tests/Cargo.toml
@@ -8,11 +8,11 @@ publish = false
 integration = []
 
 [dependencies]
-vesta-common = { path = "../vesta-common" }
+serde = { version = "1", features = ["derive"] }
 serde_json = "1"
+ureq = { version = "3", features = ["json", "rustls"] }
 tempfile = "3"
 tokio = { version = "1", features = ["rt-multi-thread", "macros", "time"] }
 tokio-tungstenite = { version = "0.29", features = ["__rustls-tls"] }
 rustls = "0.23"
 ring = "0.17"
-ureq = "3"
