+++ tests/src/types.rs
@@ -1,0 +1,70 @@
+use serde::{Deserialize, Serialize};
+
+#[derive(Debug, Clone, Serialize, Deserialize)]
+pub struct ServerConfig {
+    pub url: String,
+    pub api_key: String,
+    #[serde(default, skip_serializing_if = "Option::is_none")]
+    pub cert_fingerprint: Option<String>,
+    #[serde(default, skip_serializing_if = "Option::is_none")]
+    pub cert_pem: Option<String>,
+}
+
+#[derive(Deserialize, Serialize, Clone)]
+pub struct StatusJson {
+    pub name: String,
+    pub status: String,
+    #[serde(skip_serializing_if = "Option::is_none")]
+    pub id: Option<String>,
+    pub authenticated: bool,
+    #[serde(default)]
+    pub agent_ready: bool,
+    pub ws_port: u16,
+    pub alive: bool,
+    pub friendly_status: String,
+}
+
+#[derive(Deserialize, Serialize, Clone)]
+pub struct ListEntry {
+    pub name: String,
+    pub status: String,
+    pub authenticated: bool,
+    pub agent_ready: bool,
+    pub ws_port: u16,
+    #[serde(default)]
+    pub alive: bool,
+    #[serde(default)]
+    pub friendly_status: String,
+}
+
+#[derive(Deserialize)]
+pub struct AuthFlowResponse {
+    pub auth_url: String,
+    pub session_id: String,
+}
+
+#[derive(Debug, Deserialize)]
+pub struct StartAllResult {
+    pub name: String,
+    pub ok: bool,
+    pub error: Option<String>,
+}
+
+#[derive(Debug, Clone, Serialize, Deserialize, PartialEq)]
+#[serde(rename_all = "snake_case")]
+pub enum BackupType {
+    Manual,
+    Daily,
+    Weekly,
+    Monthly,
+    PreRestore,
+}
+
+#[derive(Debug, Clone, Serialize, Deserialize)]
+pub struct BackupInfo {
+    pub id: String,
+    pub agent_name: String,
+    pub backup_type: BackupType,
+    pub created_at: String,
+    pub size: u64,
+}
