+++ vestad/src/types.rs
@@ -1,0 +1,46 @@
+use serde::{Deserialize, Serialize};
+
+#[derive(Debug, Clone, Serialize, Deserialize, PartialEq)]
+#[serde(rename_all = "snake_case")]
+pub enum BackupType {
+    Manual,
+    Daily,
+    Weekly,
+    Monthly,
+    PreRestore,
+}
+
+impl std::fmt::Display for BackupType {
+    fn fmt(&self, f: &mut std::fmt::Formatter<'_>) -> std::fmt::Result {
+        match self {
+            Self::Manual => write!(f, "manual"),
+            Self::Daily => write!(f, "daily"),
+            Self::Weekly => write!(f, "weekly"),
+            Self::Monthly => write!(f, "monthly"),
+            Self::PreRestore => write!(f, "pre-restore"),
+        }
+    }
+}
+
+impl std::str::FromStr for BackupType {
+    type Err = String;
+    fn from_str(s: &str) -> Result<Self, Self::Err> {
+        match s {
+            "manual" => Ok(Self::Manual),
+            "daily" => Ok(Self::Daily),
+            "weekly" => Ok(Self::Weekly),
+            "monthly" => Ok(Self::Monthly),
+            "pre-restore" => Ok(Self::PreRestore),
+            other => Err(format!("unknown backup type: {}", other)),
+        }
+    }
+}
+
+#[derive(Debug, Clone, Serialize, Deserialize)]
+pub struct BackupInfo {
+    pub id: String,
+    pub agent_name: String,
+    pub backup_type: BackupType,
+    pub created_at: String,
+    pub size: u64,
+}
